:zap: Shrink Virtual_skmer from 40 to 32 bytes (32-bit last_id)
last_id only stores an index into the per-bucket (or whole-input) super-k-mer
enumeration, whose size is far below 2^32, so a 64-bit field was wasteful: it
forced 7 bytes of tail padding (24B skmer + 8B id + 1B bool -> 40B). Making it
uint32_t packs the struct to its 8-byte-alignment floor of 32 bytes, shrinking
the two ping-pong merge buffers by 20%. Kept inline (not a parallel side-array)
so the merge keeps cache locality. last_id is not serialized, so output is
byte-identical.

chr21 k31 m13, --buckets 4096: 61MB -> 54MB peak RSS (-11%), construct time
unchanged (8.07s -> 7.95s), output byte-identical.

+++ lib/include/algorithms/VirtualSkmer.hpp
@@ -149,7 +149,13 @@ inline void print_query_results(const std::vector<std::vector<uint8_t>> & result
 template<typename kuint>
 struct Virtual_skmer {
     Skmer<kuint> skmer;
-    uint64_t last_id;
+    // Index of the last enumerated super-k-mer merged into this virtual skmer.
+    // It indexes the per-bucket (or whole-input, in-RAM path) enumeration, whose
+    // size is far below 2^32, so 32 bits suffice and drop sizeof from 40 to 32
+    // bytes (the 8-byte alignment floor) — shrinking the two ping-pong merge
+    // buffers by 20%. Kept inline (not a parallel side-array) so the merge keeps
+    // its locality.
+    uint32_t last_id;
     bool expandable;
 
     Virtual_skmer() : skmer(), last_id(0), expandable(true) {}
